+++ index.php
@@ -49,6 +49,7 @@
         <li><a href="but3.php">but3 (medtog)</a></li>
         <li><a href="but4.php">but4 (alldltog)</a></li>
         <li><a href="but5.php">but5 (hotog)</a></li>
+        <li><a href="but6.php">but6 (cycle through LR modes)</a></li>
         <li><a href="xmastreeon.php">XMas tree ON (test)</a></li>
         <li><a href="xmastreeoff.php">XMas tree OFF (test)</a></li>
 		<li><a href="clearstates.php">Clear all toggle statefiles</a></li>